Create 'flipValueOf' module
Move the map of properties that require their values to be flipped, and
the value-flipping function, into their own private module.

The module includes normalization of vendor-prefixed properties,
ensuring that their values will also be flipped.
